Refactorización del proyecto, agregado del application.properties y archivo de configuración de log
