+++ src/main/java/ar/edu/davinci/dvds20221cg6/OrikaConfiguration.java
@@ -1,4 +1,4 @@
-//package ar.edu.davinci.dvds20221cg6;
+package ar.edu.davinci.dvds20221cg6;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -7,10 +7,10 @@ import org.springframework.context.annotation.Configuration;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import controller.request.PrendaInsertRequest;
-import controller.request.PrendaUpdateRequest;
-import controller.response.PrendaResponse;
-import domain.Prenda;
+import ar.edu.davinci.dvds20221cg6.controller.request.PrendaInsertRequest;
+import ar.edu.davinci.dvds20221cg6.controller.request.PrendaUpdateRequest;
+import ar.edu.davinci.dvds20221cg6.controller.response.PrendaResponse;
+import ar.edu.davinci.dvds20221cg6.domain.Prenda;
 import ma.glasnost.orika.CustomMapper;
 import ma.glasnost.orika.MapperFacade;
 import ma.glasnost.orika.MapperFactory;
@@ -19,24 +19,24 @@ import ma.glasnost.orika.impl.DefaultMapperFactory;
 
 @Configuration
 public class OrikaConfiguration {
-	
-	
+
+
 	private final Logger LOGGER = LoggerFactory.getLogger(OrikaConfiguration.class);
-	
+
 	private final ObjectMapper objectMapper;
-	
+
 	public OrikaConfiguration() {
 		objectMapper = new ObjectMapper();
 	}
-	
+
 	@Bean
 	public MapperFacade mapper() {
 		// Instanciando un mapper factory por default
 		MapperFactory mapperFactory = new DefaultMapperFactory.Builder().build();
-		
-		
+
+
 		// PRENDA
-		
+
 		mapperFactory.classMap(Prenda.class, PrendaInsertRequest.class).byDefault().register();
 		mapperFactory.classMap(Prenda.class, PrendaUpdateRequest.class).byDefault().register();
 
@@ -51,10 +51,10 @@ public class OrikaConfiguration {
 				prendaResponse.setPrecioBase(prenda.getPrecioBase());
 			}
 		}).register();
-		
+
 		// Retornameo la instancia del mapper factory
 		return mapperFactory.getMapperFacade();
 	}
-	
+
 
 }
