+++ src/main/java/ar/edu/davinci/dvds20221cg6/controller/TiendaAppRest.java
@@ -1,4 +1,4 @@
-package controller.rest;
+package ar.edu.davinci.dvds20221cg6.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -7,9 +7,8 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping(path="tienda/api")
 public abstract class TiendaAppRest {
-	
+
 	// Configurar la url de la aplicación rest
 	// http://localhost:8090/tienda/api
 
 }
-
