+++ src/main/java/ar/edu/davinci/dvds20221cg6/controller/request/PrendaUpdateRequest.java
@@ -1,4 +1,4 @@
-package controller.response;
+package ar.edu.davinci.dvds20221cg6.controller.request;
 
 import java.math.BigDecimal;
 
@@ -7,20 +7,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class PrendaResponse {
-	
-	private Long id;
-	
+public class PrendaUpdateRequest {
+
 	private String descripcion;
-	
+
 	private String tipo;
-	
+
 	private BigDecimal precioBase;
-
 }
-
