+++ src/main/java/ar/edu/davinci/dvds20221cg6/controller/response/PrendaResponse.java
@@ -1,4 +1,4 @@
-package controller.request;
+package ar.edu.davinci.dvds20221cg6.controller.response;
 
 import java.math.BigDecimal;
 
@@ -7,16 +7,19 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class PrendaInsertRequest {
+public class PrendaResponse {
+
+	private Long id;
 
 	private String descripcion;
-	
-	private String tipo;
-	
-	private BigDecimal precioBase;
-}
 
+	private String tipo;
+
+	private BigDecimal precioBase;
+
+}
