+++ src/main/java/ar/edu/davinci/dvds20221cg6/controller/rest/PrendaControllerRest.java
@@ -1,4 +1,4 @@
-package controller.rest;
+package ar.edu.davinci.dvds20221cg6.controller.rest;
 
 import java.util.List;
 
@@ -8,27 +8,25 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import controller.rest.TiendaAppRest;
-import domain.Prenda;
-import ma.glasnost.orika.MapperFacade;
-import service.PrendaService;
+import ar.edu.davinci.dvds20221cg6.controller.TiendaAppRest;
+import ar.edu.davinci.dvds20221cg6.domain.Prenda;
+import ar.edu.davinci.dvds20221cg6.service.PrendaService;
 
 @RestController
 public class PrendaControllerRest extends TiendaAppRest {
 
 	private final Logger LOGGER = LoggerFactory.getLogger(PrendaControllerRest.class);
-	
+
 	@Autowired
 	private PrendaService service;
-	
+
 	/**
 	 * Listar Prendas
 	 */
 	@GetMapping(path = "/prendas/all")
-	public List<Prenda> getList(){
+	public List<Prenda> getList() {
 		LOGGER.info("Listado de todas las prendas");
 		return service.list();
 	}
-	
-}
 
+}
