+++ src/main/java/ar/edu/davinci/dvds20221cg6/domain/Prenda.java
@@ -1,4 +1,4 @@
-package domain;
+package ar.edu.davinci.dvds20221cg6.domain;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
@@ -30,9 +30,9 @@ import lombok.NoArgsConstructor;
 @Builder
 
 public class Prenda implements Serializable {
-	
+
 	private static final long serialVersionUID = 8791033026571491633L;
-	
+
 	// Configurar por JPA cual el PK de la tabla prendas
 	@Id
 	// Configurar la estragia de generación de los ids por JPA
@@ -41,16 +41,15 @@ public class Prenda implements Serializable {
 	// Configuramos por JPA el nombre de la columna
 	@Column(name = "prd_id")
 	private Long id;
-	
+
 	@Column(name = "prd_descripcion", nullable = false)
 	private String descripcion;
-	
+
 	@Column(name = "prd_tipo_prenda")
 	@Enumerated(EnumType.STRING)
 	private TipoPrenda tipo;
-	
+
 	@Column(name = "prd_precio_base")
 	private BigDecimal precioBase;
 
 }
-
