+++ src/main/java/ar/edu/davinci/dvds20221cg6/exception/BusinessException.java
@@ -1,10 +1,9 @@
-package exception;
+package ar.edu.davinci.dvds20221cg6.exception;
 
 public class BusinessException extends Exception {
-	
+
 	public BusinessException(String message) {
 		super(message);
 	}
 
 }
-
