+++ src/main/java/ar/edu/davinci/dvds20221cg6/repository/PrendaRepository.java
@@ -1,13 +1,12 @@
-package repository;
+package ar.edu.davinci.dvds20221cg6.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import domain.Prenda;
+import ar.edu.davinci.dvds20221cg6.domain.Prenda;
+
 
 @Repository
 public interface PrendaRepository extends JpaRepository<Prenda, Long> {
 
 }
-
-
