+++ src/main/java/ar/edu/davinci/dvds20221cg6/service/PrendaService.java
@@ -1,16 +1,17 @@
-package service;
+package ar.edu.davinci.dvds20221cg6.service;
 
 import java.util.List;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import domain.Prenda;
-import exception.BusinessException;
+import ar.edu.davinci.dvds20221cg6.domain.Prenda;
+import ar.edu.davinci.dvds20221cg6.exception.BusinessException;
+
 
 
 public interface PrendaService {
-	
+
 	// Métodos de negocio sobre la entida Prenda
 
 	// Métodos de creación, modificación y eliminación de una prenda
@@ -21,11 +22,11 @@ public interface PrendaService {
 
 	// Métodos de búsqueda
 	Prenda findById(Long id) throws BusinessException;
-	
+
 	// Métodos de listado
 	List<Prenda> list();
 	Page<Prenda> list(Pageable pegeable);
-	
+
 	// Método de contar la cantidad de registros
 	long count();
