+++ src/test/resources/application.properties
@@ -1,0 +1,32 @@
+# spring.datasource.x
+spring.datasource.url=jdbc:h2:mem:db;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
+spring.datasource.driver-class-name=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=sa
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.datasource.initialization-mode=always
+spring.jpa.hibernate.ddl-auto=none
+
+
+# hibernate.X
+hibernate.dialect=org.hibernate.dialect.H2Dialect
+hibernate.show_sql=true
+#hibernate.hbm2ddl.auto=create-drop
+hibernate.cache.use_second_level_cache=true
+hibernate.cache.use_query_cache=true
+hibernate.cache.region.factory_class=org.hibernate.cache.ehcache.EhCacheRegionFactory
+
+
+spring.datasource.schema=classpath:/data/h2db/schema.sql
+spring.datasource.data=classpath:/data/h2db/data.sql
+
+spring.h2.console.enabled=true
+
+
+# Show SQL
+spring.jpa.show-sql=false
+spring.jpa.properties.hibernate.format_sql=true
+
+#Login sentences
+logging.level.org.hibernate.SQL=DEBUG
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
