+++ src/test/resources/data/h2db/data.sql
@@ -1,0 +1,10 @@
+INSERT INTO PRENDAS
+(prd_id,
+prd_descripcion,
+prd_precio_base,
+prd_tipo_prenda)
+VALUES
+(1, 'Camisa Celeste', 10.24, 'CAMISA'),
+(2, 'Camisa Blanca', 100.50, 'CAMISA'),
+(3, 'Saco Vestir', 102.4, 'SACO'),
+(4, 'Pantalon Gabardina Beige', 1004, 'PANTALON');
